Refactor: clean up malloc/realloc/free/calloc
Added memcached_set_memory_allocators and memcached_get_memory_allocators
to set and get the memory allocators used by memcached. The old versions set
by the memcached_set_callback function is deprecated and will be removed in
future versions.

Internally we will now only use the call_malloc function pointer and don't
test if it is set before calling (it is initialized to a libc-version)